# swagger-spec.json
{"openapi":"3.0.0","paths":{"/api/auth/register":{"post":{"operationId":"AuthController_register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccountForPost"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccountForFull"}}}}},"tags":["auth"]}},"/api/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccountForLogin"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccountForLoginResponse"}}}}},"tags":["auth"]}},"/api/auth/profile":{"get":{"operationId":"AuthController_getProfile","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccountForFull"}}}}},"tags":["auth"],"security":[{"Authorization":[]}]}},"/api/post":{"post":{"operationId":"PostController_createPost","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostForCreate"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostForResponse"}}}}},"tags":["post"],"security":[{"Authorization":[]}]},"patch":{"operationId":"PostController_updatePost","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostForUpdate"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostForResponse"}}}}},"tags":["post"],"security":[{"Authorization":[]}]}},"/api/post/{id}":{"delete":{"operationId":"PostController_deletePost","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"delete ok","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostForResponse"}}}}},"tags":["post"],"security":[{"Authorization":[]}]}},"/api/post/valid-post":{"get":{"operationId":"PostController_getValidPostByUser","parameters":[{"name":"pageNo","required":false,"in":"query","schema":{"type":"number"}},{"name":"limit","required":false,"in":"query","schema":{"type":"number"}},{"name":"totalPage","required":false,"in":"query","schema":{"type":"number"}},{"name":"totalRecord","required":false,"in":"query","schema":{"type":"number"}},{"name":"sortBy","required":false,"in":"query","schema":{"type":"string"}},{"name":"orderBy","required":false,"in":"query","schema":{"type":"string"}},{"required":false,"name":"pageNo","in":"query","schema":{"type":"number"}},{"required":false,"name":"limit","in":"query","schema":{"type":"number"}},{"required":false,"name":"totalPage","in":"query","schema":{"type":"number"}},{"required":false,"name":"totalRecord","in":"query","schema":{"type":"number"}},{"required":false,"name":"sortBy","in":"query","schema":{"type":"string"}},{"required":false,"name":"orderBy","in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostForFullResponse"}}}}},"tags":["post"],"security":[{"Authorization":[]},{"Authorization":[]}]}},"/api/post/get-all-post":{"get":{"operationId":"PostController_getAllPost","parameters":[{"name":"limit","required":true,"in":"query","schema":{"type":"number"}},{"name":"pageNo","required":true,"in":"query","schema":{"type":"number"}},{"name":"sortBy","required":true,"in":"query","schema":{"type":"string"}},{"name":"orderBy","required":true,"in":"query","schema":{"type":"string"}},{"name":"contentTextKey","required":true,"in":"query","schema":{"type":"string"}},{"name":"nameAccountKey","required":true,"in":"query","schema":{"type":"string"}},{"name":"emailAccountKey","required":true,"in":"query","schema":{"type":"string"}},{"name":"createdDateFrom","required":true,"in":"query","schema":{"type":"string"}},{"name":"createdDateTo","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostForFullResponse"}}}}},"tags":["post"],"security":[{"Authorization":[]}]}},"/api/user/update-account-admin":{"patch":{"operationId":"UserController_updateAdministratorAccount","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdminAccountForPut"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdminAccountForResponse"}}}}},"tags":["user"],"security":[{"Authorization":[]}]}},"/api/user/suggest-follow":{"get":{"operationId":"UserController_suggestFollowForAccount","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserForResponse"}}}}}},"tags":["user"],"security":[{"Authorization":[]}]}},"/api/user/{id}":{"patch":{"operationId":"UserController_updateUser","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserForUpdate"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserForResponse"}}}}},"tags":["user"],"security":[{"Authorization":[]}]}},"/api/roles":{"get":{"operationId":"RoleController_getAllRoles","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RoleForGet"}}}}}},"tags":["Role"],"security":[{"Authorization":[]}]},"post":{"operationId":"RoleController_addOneRole","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoleForPost"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoleForGet"}}}}},"tags":["Role"],"security":[{"Authorization":[]}]}},"/api/roles/{id}":{"patch":{"operationId":"RoleController_updateOneRole","parameters":[{"name":"id","required":true,"in":"path","description":"this is id of one role","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoleForPost"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoleForGet"}}}}},"tags":["Role"],"security":[{"Authorization":[]}]},"delete":{"operationId":"RoleController_deleteOneRole","parameters":[{"name":"id","required":true,"in":"path","description":"this is id of one role","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoleForGet"}}}}},"tags":["Role"],"security":[{"Authorization":[]}]}},"/api/feature":{"post":{"operationId":"FeatureController_createNewFeature","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FeatureForPost"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FeatureForGet"}}}}},"tags":["feature"]},"get":{"operationId":"FeatureController_getAllFeature","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/FeatureForGet"}}}}}},"tags":["feature"]}},"/api/feature/{id}":{"patch":{"operationId":"FeatureController_updateFeature","parameters":[{"name":"id","required":true,"in":"query","description":"id of feature","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FeatureForUpdate"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FeatureForGet"}}}}},"tags":["feature"]},"delete":{"operationId":"FeatureController_deleteFeature","parameters":[{"name":"id","required":true,"in":"query","description":"id of feature","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FeatureForGet"}}}}},"tags":["feature"]}},"/api/hot-content":{"post":{"operationId":"HotContentController_createNewHotContent","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HotContentForPost"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HotContentForResponse"}}}}},"tags":["hot-content"]},"get":{"operationId":"HotContentController_getAllHotContent","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/HotContentForResponse"}}}}}},"tags":["hot-content"]}},"/api/hot-content/{id}":{"patch":{"operationId":"HotContentController_updateHotContent","parameters":[{"name":"id","required":true,"in":"query","description":"id of hot content","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HotContentForUpdate"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HotContentForResponse"}}}}},"tags":["hot-content"]},"delete":{"operationId":"HotContentController_deleteHotContent","parameters":[{"name":"id","required":true,"in":"query","description":"id of HotContent","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HotContentForResponse"}}}}},"tags":["hot-content"]}},"/api/tab-menu":{"post":{"operationId":"TabMenuController_createNewTabMenu","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TabMenuForPost"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TabMenuForGet"}}}}},"tags":["tab-menu"]},"get":{"operationId":"TabMenuController_getAllTabMenu","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TabMenuForGet"}}}}}},"tags":["tab-menu"]}},"/api/tab-menu/{id}":{"patch":{"operationId":"TabMenuController_updateFeature","parameters":[{"name":"id","required":true,"in":"query","description":"id of hot tab menu","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TabMenuForUpdate"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TabMenuForGet"}}}}},"tags":["tab-menu"]},"delete":{"operationId":"TabMenuController_deleteTabMenu","parameters":[{"name":"id","required":true,"in":"query","description":"id of hot tab menu","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TabMenuForGet"}}}}},"tags":["tab-menu"]}},"/api/room-message":{"post":{"operationId":"RoomMessageController_createNewRoomMessage","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoomMessageForPost"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoomMessageForGet"}}}}},"tags":["room-message"]},"get":{"operationId":"RoomMessageController_getAllValidRoomMessage","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RoomMessageForGet"}}}}}},"tags":["room-message"]}},"/api/likes":{"post":{"operationId":"LikeController_updateStatusLikes","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LikeForPost"}}}},"responses":{"201":{"description":""}},"security":[{"Authorization":[]}]}},"/api/comment":{"post":{"operationId":"CommentController_createComment","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommentForCreate"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommentForResponse"}}}}},"tags":["comment"],"security":[{"Authorization":[]}]}},"/api/comment/{id}":{"patch":{"operationId":"CommentController_updateComment","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommentForUpdate"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommentForResponse"}}}}},"tags":["comment"],"security":[{"Authorization":[]}]},"delete":{"operationId":"CommentController_deleteComment","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommentForResponse"}}}}},"tags":["comment"],"security":[{"Authorization":[]}]}},"/api/file/post/image":{"post":{"operationId":"FileController_uploadPostImage","parameters":[],"responses":{"201":{"description":""}}}},"/api/file/avata/image":{"post":{"operationId":"FileController_uploadAvataImage","parameters":[],"responses":{"201":{"description":""}}}},"/api/follow":{"post":{"operationId":"FollowController_follow","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FollowForCreate"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FollowForGet"}}}}},"tags":["follows"]}},"/api/follow/{id}":{"delete":{"operationId":"FollowController_deleteRequestFollow","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["follows"]}},"/api/follow/accept/{id}":{"post":{"operationId":"FollowController_acceptRequestFollow","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FollowForCreate"}}}},"responses":{"201":{"description":""}},"tags":["follows"]}},"/api/follow/unfollow/{id}":{"delete":{"operationId":"FollowController_unFollow","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["follows"]}},"/api/follow/followings":{"get":{"operationId":"FollowController_getAllFollowingOfAccount","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserForResponse"}}}}}},"tags":["follows"]}},"/api/follow/followers":{"get":{"operationId":"FollowController_getAllFollowerOfAccount","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserForResponse"}}}}}},"tags":["follows"]}},"/api/follow/request-followers":{"get":{"operationId":"FollowController_getAllRequestFollowers","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RequestFollowForResponse"}}}}}},"tags":["follows"]}}},"info":{"title":"mental heath","description":"Một dự án phát triển bởi những developers đẳng cấp","version":"1.0","contact":{}},"tags":[{"name":"mental heath","description":""}],"servers":[],"components":{"securitySchemes":{"Authorization":{"scheme":"Bearer","bearerFormat":"Bearer","description":"[just text field] Please enter token in following format: Bearer <JWT>","name":"Authorization","type":"http","in":"Header"}},"schemas":{"AccountForPost":{"type":"object","properties":{"fullName":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"}},"required":["fullName","email","password"]},"RoleForGet":{"type":"object","properties":{"id":{"type":"string"},"nameRole":{"type":"string"},"descriptionRole":{"type":"string"},"created_at":{"format":"date-time","type":"string"},"updated_at":{"format":"date-time","type":"string"}},"required":["id","nameRole","descriptionRole","created_at","updated_at"]},"AccountForFull":{"type":"object","properties":{"id":{"type":"string"},"fullName":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"},"phone":{"type":"string"},"aboutMe":{"type":"string"},"nickName":{"type":"string"},"birth":{"format":"date-time","type":"string"},"address":{"type":"string"},"refreshTokenJWT":{"type":"string"},"expriedTokenJWT":{"format":"date-time","type":"string"},"roles":{"type":"array","items":{"$ref":"#/components/schemas/RoleForGet"}},"created_at":{"format":"date-time","type":"string"},"updated_at":{"format":"date-time","type":"string"}},"required":["id","fullName","email","password","phone","aboutMe","nickName","birth","address","refreshTokenJWT","expriedTokenJWT","roles","created_at","updated_at"]},"AccountForLogin":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]},"AccountForLoginResponse":{"type":"object","properties":{"idAccount":{"type":"string"},"access_token":{"type":"string"}},"required":["idAccount","access_token"]},"PostForCreate":{"type":"object","properties":{"contentText":{"type":"string"},"accountId":{"type":"string"}},"required":["contentText","accountId"]},"UserForResponse":{"type":"object","properties":{"id":{"type":"string"},"fullName":{"type":"string"},"phone":{"type":"string"},"aboutMe":{"type":"string"},"nickName":{"type":"string"},"birth":{"format":"date-time","type":"string"},"address":{"type":"string"}},"required":["id","fullName","phone","aboutMe","nickName","birth","address"]},"PostForResponse":{"type":"object","properties":{"id":{"type":"string"},"contentText":{"type":"string"},"accountId":{"type":"string"},"account":{"$ref":"#/components/schemas/UserForResponse"},"created_at":{"format":"date-time","type":"string"},"updated_at":{"format":"date-time","type":"string"},"totalReaction":{"type":"number"},"totalComment":{"type":"number"},"totalShare":{"type":"number"}},"required":["id","contentText","accountId","account","created_at","updated_at","totalReaction","totalComment","totalShare"]},"PostForUpdate":{"type":"object","properties":{"id":{"type":"string"},"contentText":{"type":"string"},"accountId":{"type":"string"}},"required":["id","contentText","accountId"]},"PaginationAndFilter":{"type":"object","properties":{"pageNo":{"type":"number"},"limit":{"type":"number"},"totalPage":{"type":"number"},"totalRecord":{"type":"number"},"sortBy":{"type":"string"},"orderBy":{"type":"string"}}},"PostForFullResponse":{"type":"object","properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/PostForResponse"}},"pagination":{"$ref":"#/components/schemas/PaginationAndFilter"}},"required":["data"]},"AdminAccountForPut":{"type":"object","properties":{"idAccount":{"type":"string"},"idRoleAdmin":{"type":"string"}},"required":["idAccount","idRoleAdmin"]},"AdminAccountForResponse":{"type":"object","properties":{"id":{"type":"string"},"fullName":{"type":"string"},"roles":{"type":"array","items":{"$ref":"#/components/schemas/RoleForGet"}},"phone":{"type":"string"},"updated_at":{"format":"date-time","type":"string"}},"required":["id","fullName","roles","phone","updated_at"]},"UserForUpdate":{"type":"object","properties":{"id":{"type":"string"},"fullName":{"type":"string"},"phone":{"type":"string"},"aboutMe":{"type":"string"},"nickName":{"type":"string"},"birth":{"format":"date-time","type":"string"},"address":{"type":"string"}},"required":["id","fullName","phone","aboutMe","nickName","birth","address"]},"RoleForPost":{"type":"object","properties":{"nameRole":{"type":"string"},"descriptionRole":{"type":"string"}},"required":["nameRole","descriptionRole"]},"FeatureForPost":{"type":"object","properties":{"name":{"type":"string"},"thumbnailFileName":{"type":"string"},"url":{"type":"string"}}},"FeatureForGet":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"thumbnailFileName":{"type":"string"},"url":{"type":"string"},"created_at":{"format":"date-time","type":"string"},"updated_at":{"format":"date-time","type":"string"}}},"FeatureForUpdate":{"type":"object","properties":{"name":{"type":"string"},"thumbnailFileName":{"type":"string"},"url":{"type":"string"}}},"HotContentForPost":{"type":"object","properties":{"title":{"type":"string"},"thumbnailFileName":{"type":"string"},"url":{"type":"string"}}},"HotContentForResponse":{"type":"object","properties":{"id":{"type":"string"},"title":{"type":"string"},"thumbnailFileName":{"type":"string"},"url":{"type":"string"}}},"HotContentForUpdate":{"type":"object","properties":{"title":{"type":"string"},"thumbnailFileName":{"type":"string"},"url":{"type":"string"}}},"TabMenuForPost":{"type":"object","properties":{"name":{"type":"string"},"iconUrl":{"type":"string"},"url":{"type":"string"}}},"TabMenuForGet":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"iconUrl":{"type":"string"},"url":{"type":"string"},"created_at":{"format":"date-time","type":"string"},"updated_at":{"format":"date-time","type":"string"}}},"TabMenuForUpdate":{"type":"object","properties":{"name":{"type":"string"},"iconUrl":{"type":"string"},"url":{"type":"string"}}},"RoomMessageForPost":{"type":"object","properties":{"accountInRoom":{"type":"array","items":{"type":"string"}}},"required":["accountInRoom"]},"RoomMessageForGet":{"type":"object","properties":{"id":{"type":"string"},"created_at":{"format":"date-time","type":"string"},"updated_at":{"format":"date-time","type":"string"},"accountInRoom":{"type":"array","items":{"type":"string"}},"nameRoom":{"type":"string"}},"required":["id","created_at","updated_at","accountInRoom","nameRoom"]},"LikeForPost":{"type":"object","properties":{"accountId":{"type":"string"},"postId":{"type":"string"}},"required":["accountId","postId"]},"CommentForCreate":{"type":"object","properties":{"contentCmt":{"type":"string"},"accountId":{"type":"string"},"postId":{"type":"string"}},"required":["contentCmt","accountId","postId"]},"CommentForResponse":{"type":"object","properties":{"id":{"type":"string"},"contentCmt":{"type":"string"},"accountId":{"type":"string"},"created_at":{"type":"object"},"updated_at":{"type":"object"}},"required":["id","contentCmt","accountId","created_at","updated_at"]},"CommentForUpdate":{"type":"object","properties":{"contentCmt":{"type":"string"},"accountId":{"type":"string"}},"required":["contentCmt","accountId"]},"FollowForCreate":{"type":"object","properties":{"senderId":{"type":"string"},"reciverId":{"type":"string"}},"required":["senderId","reciverId"]},"FollowForGet":{"type":"object","properties":{"id":{"type":"string"},"senderId":{"type":"string"},"sender":{"$ref":"#/components/schemas/UserForResponse"},"reciver":{"$ref":"#/components/schemas/UserForResponse"},"reciverId":{"type":"string"},"updated_at":{"format":"date-time","type":"string"},"created_at":{"format":"date-time","type":"string"}},"required":["id","senderId","sender","reciver","reciverId","updated_at","created_at"]},"RequestFollowForResponse":{"type":"object","properties":{"id":{"type":"string"},"sender":{"$ref":"#/components/schemas/UserForResponse"},"created_at":{"type":"object"},"updated_at":{"type":"object"}},"required":["id","sender","created_at","updated_at"]}}}}